requirements.txt
absl-py==2.3.0
alembic==1.15.2
altgraph==0.17.4
astunparse==1.6.3
backports.tarfile==1.2.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
docutils==0.21.2
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
gast==0.6.0
git-filter-repo==2.47.0
google-auth==2.40.1
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
greenlet==3.2.1
grpcio==1.72.1
h5py==3.13.0
id==1.5.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
keras==3.10.0
keyring==25.6.0
libclang==18.1.1
macholib==1.16.3
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.7.0
namex==0.1.0
nh3==0.2.21
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.5.5.64
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
protobuf==4.25.8
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyJWT==2.10.1
pyproject_hooks==1.2.0
python-dotenv==1.1.0
readme_renderer==44.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.40
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-estimator==2.15.0
tensorflow-intel==0.0.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tomli==2.2.1
twine==6.1.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.22.0
